GlossaryTerm, refactor add_pattern_condition
