+++ app/classes/query/collection_number_base.rb
@@ -45,13 +45,6 @@ class Query::CollectionNumberBase < Query::Base
     add_join(:observation_collection_numbers)
   end
 
-  def add_pattern_condition
-    return if params[:pattern].blank?
-
-    @title_tag = :query_title_pattern_search
-    add_search_condition(search_fields, params[:pattern])
-  end
-
   def search_fields
     "CONCAT(" \
       "collection_numbers.name," \
