+++ app/classes/query/glossary_term_base.rb
@@ -12,16 +12,25 @@ module Query
         updated_at?: [:time],
         users?: [User],
         name_has?: :string,
-        description_has?: :string
+        description_has?: :string,
+        pattern?: :string
       )
     end
 
     def initialize_flavor
       add_owner_and_time_stamp_conditions("glossary_terms")
+      add_pattern_condition
       add_search_condition("glossary_terms.name", params[:name_has])
       add_search_condition("glossary_terms.description",
                            params[:description_has])
       super
+    end
+
+    def search_fields
+      "CONCAT(" \
+        "glossary_terms.name," \
+        "COALESCE(glossary_terms.description,'')" \
+        ")"
     end
 
     def self.default_order
