+++ app/classes/query/herbarium_record_base.rb
@@ -55,13 +55,6 @@ class Query::HerbariumRecordBase < Query::Base
     where << "herbarium_records.herbarium_id = '#{herbarium.id}'"
   end
 
-  def add_pattern_condition
-    return if params[:pattern].blank?
-
-    @title_tag = :query_title_pattern_search
-    add_search_condition(search_fields, params[:pattern])
-  end
-
   def search_fields
     "CONCAT(" \
       "herbarium_records.initial_det," \
